+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 - MapMarker: Added function for rotation in 3D.
 - Reusable component SimpleButton.
 - Added a config.json configuration in addition to highlighting features: 'treeHighlightedFeatures'. If features are highlighted with the "List" or "Select Features" tool with "Zoom to this Feature" or via url parameter, then a layer with these features is selectable in the menu tree.
+- Filter:
+  - STA: There is now the possibility to filter SensorThingsAPI layers.
 
 ### Changed
 - OAF: